requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==2.3.7
bcrypt==4.0.1
mercadopago==2.2.0
python-dotenv==1.0.0
Pillow==10.0.0
email-validator==2.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23 